# Performance improvements for build_tsibble