--- a/R/as-tsibble.R
+++ b/R/as-tsibble.R
@@ -283,7 +283,7 @@ build_tsibble_meta <- function(
 ) {
   stopifnot(is_string(index) && is_string(index2))
   stopifnot(!is_null(ordered))
-  stopifnot(tibble::is_tibble(x))
+  stopifnot(is.data.frame(x))
   tbl <- x
   attr(index, "ordered") <- ordered