--- a/R/deprecated.R
+++ b/R/deprecated.R
@@ -68,6 +68,9 @@ id <- function(...) {
 
 use_id <- function(x, key) {
   key_quo <- enquo(key)
+  if(quo_is_null(key_quo)){
+    return(character())
+  }
   if (quo_is_call(key_quo)) {
     call_fn <- call_name(key_quo)
     if (call_fn == "id") {