@earowang 2019-06-05
I'd suggest to use `new_tsibble()` where possible for your performance gain, which does very minimum check. But for manipulation verbs such as `mutate()`, use `build_tsibble()` to catch the corner cases.