@mitchelloharawild 2019-06-06
> I'd suggest to use `new_tsibble()` where possible for your performance gain, which does very minimum check. But for manipulation verbs such as `mutate()`, use `build_tsibble()` to catch the corner cases.

`new_tsibble` is used to create extension objects (like `fbl_ts`) - I'm looking to create a `tbl_ts` without checks and computation of `build_tsibble()`. Should I use `new_tibble()`?