@earowang 2019-06-06
> `new_tsibble` is used to create extension objects (like `fbl_ts`) - I'm looking to create a `tbl_ts` without checks and computation of `build_tsibble()`. Should I use `new_tibble()`?

I was suggesting to have an internal function like `tsibble_restore()` that wraps `new_tsibble()`. `new_*` is not only for extensions but also avoid checks like `new_grouped_df()`.

I could export `build_tsibble_meta()` if needed.