@mitchelloharawild 2019-06-06
Exporting `build_tsibble_meta` would help speed things up, but even that has some unnecessary overhead for what I need. I'm not sure how using `new_tsibble()` can be used to create a tsibble efficiently - the first thing it does is check that the input is a tsibble.